+++ rio_tiler/reader.py
@@ -298,7 +298,6 @@ def read(
             data,
             bounds=out_bounds,
             crs=dataset.crs,
-            band_names=[f"b{idx}" for idx in indexes],
             band_descriptions=[
                 dataset.descriptions[ix - 1] or f"b{idx}" for idx in indexes
             ],
@@ -511,7 +510,6 @@ def part(
             img.array[:, padding:-padding, padding:-padding],
             bounds=bounds,
             crs=img.crs,
-            band_names=img.band_names,
             band_descriptions=img.band_descriptions,
             nodata=img.nodata,
             scales=img.scales,
@@ -651,7 +649,6 @@ def point(
 
         return PointData(
             img.array[:, 0, 0],
-            band_names=img.band_names,
             band_descriptions=img.band_descriptions,
             coordinates=coordinates,
             crs=coord_crs,