+++ tests/test_io_xarray.py
@@ -76,9 +76,11 @@ def test_xarray_reader():
 
         stats = dst.statistics(indexes=1)
         assert list(stats) == ["b1"]
+        assert stats["b1"].description == "2022-01-01T00:00:00.000000000"
 
         stats = dst.statistics(indexes=2)
-        assert list(stats) == ["b2"]
+        assert list(stats) == ["b1"]
+        assert stats["b1"].description == "2022-01-02T00:00:00.000000000"
 
         stats = dst.statistics(indexes=(1, 2))
         assert list(stats) == [
@@ -97,10 +99,6 @@ def test_xarray_reader():
         assert img.width == 256
         assert img.height == 256
         assert img.array.dtype == numpy.float32
-        assert img.band_names == [
-            "b1",
-            "b2",
-        ]
         assert img.band_descriptions == [
             "2022-01-01T00:00:00.000000000",
             "2022-01-02T00:00:00.000000000",
@@ -109,9 +107,6 @@ def test_xarray_reader():
 
         img = dst.tile(0, 0, 0, indexes=2)
         assert img.count == 1
-        assert img.band_names == [
-            "b2",
-        ]
         assert img.band_descriptions == [
             "2022-01-02T00:00:00.000000000",
         ]
@@ -141,7 +136,6 @@ def test_xarray_reader():
         img = dst.part((-160, -80, 160, 80), indexes=1)
         assert img.crs == "epsg:4326"
         assert img.count == 1
-        assert img.band_names == ["b1"]
         assert img.band_descriptions == ["2022-01-01T00:00:00.000000000"]
         assert img.array.shape == (1, 32, 32)
         assert img.array.dtype == numpy.float32
@@ -152,7 +146,6 @@ def test_xarray_reader():
         img = dst.part((-160, -80, 160, 80), dst_crs="epsg:3857", indexes=1)
         assert img.crs == "epsg:3857"
         assert img.count == 1
-        assert img.band_names == ["b1"]
         assert img.band_descriptions == ["2022-01-01T00:00:00.000000000"]
         assert img.array.shape == (1, 29, 33)
 
@@ -171,10 +164,6 @@ def test_xarray_reader():
         assert img.crs == "epsg:4326"
         assert img.count == 2
         assert img.array.dtype == numpy.float32
-        assert img.band_names == [
-            "b1",
-            "b2",
-        ]
         assert img.band_descriptions == [
             "2022-01-01T00:00:00.000000000",
             "2022-01-02T00:00:00.000000000",
@@ -187,14 +176,12 @@ def test_xarray_reader():
         img = dst.preview(indexes=1)
         assert img.crs == "epsg:4326"
         assert img.count == 1
-        assert img.band_names == ["b1"]
         assert img.band_descriptions == ["2022-01-01T00:00:00.000000000"]
         assert img.array.shape == (1, 33, 35)
 
         img = dst.preview(dst_crs="epsg:3857", indexes=2)
         assert img.crs == "epsg:3857"
         assert img.count == 1
-        assert img.band_names == ["b2"]
         assert img.band_descriptions == ["2022-01-02T00:00:00.000000000"]
         assert img.array.shape == (1, 32, 36)
 
@@ -212,10 +199,6 @@ def test_xarray_reader():
 
         pt = dst.point(0, 0)
         assert pt.count == 2
-        assert pt.band_names == [
-            "b1",
-            "b2",
-        ]
         assert pt.band_descriptions == [
             "2022-01-01T00:00:00.000000000",
             "2022-01-02T00:00:00.000000000",
@@ -238,7 +221,6 @@ def test_xarray_reader():
 
         pt = dst.point(0, 0, indexes=2)
         assert pt.count == 1
-        assert pt.band_names == ["b2"]
         assert pt.band_descriptions == ["2022-01-02T00:00:00.000000000"]
         assert pt.coordinates
         xys = [[0, 2.499], [0, 2.501], [-4.999, 0], [-5.001, 0], [-170, 80]]
@@ -273,10 +255,6 @@ def test_xarray_reader():
         img = dst.feature(feat)
         assert img.count == 2
         assert img.array.dtype == numpy.float32
-        assert img.band_names == [
-            "b1",
-            "b2",
-        ]
         assert img.band_descriptions == [
             "2022-01-01T00:00:00.000000000",
             "2022-01-02T00:00:00.000000000",
@@ -289,13 +267,11 @@ def test_xarray_reader():
 
         img = dst.feature(feat, indexes=2)
         assert img.count == 1
-        assert img.band_names == ["b2"]
         assert img.band_descriptions == ["2022-01-02T00:00:00.000000000"]
         assert img.array.shape == (1, 24, 31)
 
         img = dst.feature(feat, dst_crs="epsg:3857", indexes=1)
         assert img.count == 1
-        assert img.band_names == ["b1"]
         assert img.band_descriptions == ["2022-01-01T00:00:00.000000000"]
         assert img.crs == "epsg:3857"
         assert img.array.shape == (1, 19, 34)
@@ -322,7 +298,6 @@ def test_xarray_reader():
         assert img.count == 1
         assert img.width == 256
         assert img.height == 256
-        assert img.band_names == ["b1"]
         assert img.band_descriptions == ["2022-01-01T00:00:00.000000000"]
         assert img.dataset_statistics == ((arr.min(), arr.max()),)
 
@@ -705,7 +680,6 @@ def test_xarray_reader_no_dims():
         assert img.count == 1
         assert img.width == 256
         assert img.height == 256
-        assert img.band_names == ["b1"]
         assert img.band_descriptions == ["b1"]
         assert img.dataset_statistics == ((arr.min(), arr.max()),)
 
@@ -713,13 +687,11 @@ def test_xarray_reader_no_dims():
         assert img.count == 1
         assert img.width == 32
         assert img.height == 32
-        assert img.band_names == ["b1"]
         assert img.band_descriptions == ["b1"]
         assert img.dataset_statistics == ((arr.min(), arr.max()),)
 
         pt = dst.point(0, 0)
         assert pt.count == 1
-        assert pt.band_names == ["b1"]
         assert img.band_descriptions == ["b1"]
         assert pt.coordinates
         xys = [[0, 2.499], [0, 2.501], [-4.999, 0], [-5.001, 0], [-170, 80]]
@@ -752,7 +724,6 @@ def test_xarray_reader_dims_order():
         info = dst.info()
         assert info.band_descriptions == [("b1", "2022-01-01T00:00:00.000000000")]
         pt = dst.point(0, 0)
-        assert pt.band_names == ["b1"]
         assert pt.band_descriptions == ["2022-01-01T00:00:00.000000000"]